// web-mart-front-end-main/src/store/auth/auth.enum.ts
enum AuthActionTypeEnum {
  AUTH_PENDING = "AUTH_PENDING",
  AUTH_SUCCESS = "AUTH_SUCCESS",
  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  SIGNUP_SUCCESS = "SIGNUP_SUCCESS",
  LOGOUT = "LOGOUT",
  FORGOT_PASSWORD = "FORGOT_PASSWORD",
  RESET_FORGOT_PASSWORD = "RESET_FORGOT_PASSWORD",
  CHANGE_PASSWORD = "CHANGE_PASSWORD",
}

export default AuthActionTypeEnum;